+++ packages/geoinfo/client/views/panels.js
@@ -1,5 +1,5 @@
 
-Template.panelPlaceEdit_geoinfo.onRendered(function() {
+/*Template.panelPlaceEdit_geoinfo.onRendered(function() {
 	
 	var panel$ = $(this.firstNode).parents('.panel-body'),
 		input$ = panel$.find('.input-editren'),
@@ -7,4 +7,16 @@ Template.panelPlaceEdit_geoinfo.onRendered(function() {
 
 	if(input$.val()==='' && near!=='')
 		input$.val(near);
+});
+
+*/
+Template.panelPlaceEdit_geoinfo.events({
+	'click .geoinfo-nametips .btn': function(e,tmpl) {
+		
+		var panel$ = $(tmpl.firstNode).parents('.panel-body'),
+			input$ = panel$.find('.input-editren'),
+			name = $(e.target).text();
+
+		input$.val(name);
+	}
 });