+++ packages/geoinfo/i18n/en.js
@@ -15,5 +15,8 @@ i18n.map('en', {
 	label_sunrise: "Sunrise",
 	label_sunset: "Sunset",
 
+	label_nametips: "Name tips by locality",
+
 	btn_geosearch: "Geography Info",
+
 });