+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 - plugin geoinfo suggest name in panel edit
 - new method K.Profile.setMobile()
 - map gps fix gps user
+- plugin geoinfo in panelEdit Name tips by locality
 
 ## v1.5.7
 - plugin admin new page admin users